requirements.txt
# Computer Vision & Detection
ultralytics>=8.0.0
torch>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# AI & LLM
openai>=1.0.0
groq

# Utilities
python-dotenv>=1.0.0